Update GitHub login button label
